NNNN/requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
faiss-gpu>=1.7.2
torch>=2.0.0
transformers>=4.36.0
sentence-transformers>=2.2.2
hanlp>=2.1.0
python-docx>=0.8.11
PyPDF2>=3.0.0
numpy>=1.24.3
auto-gptq>=0.5.0
optimum>=1.16.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
flask>=2.3.0
flask-cors>=4.0.0 